requirements.txt
flask==2.3.3
python-chess==1.999
flask-socketio==5.3.4
python-engineio==4.5.1
python-socketio==5.8.0 